Refactor: move wallet utils to utils (#2047)
