+++ src/components/tx/SignOrExecuteForm/hooks.test.ts
@@ -4,7 +4,7 @@ import { type SafeInfo } from '@safe-global/safe-gateway-typescript-sdk'
 import { type ConnectedWallet } from '@/services/onboard'
 import * as useSafeInfoHook from '@/hooks/useSafeInfo'
 import * as wallet from '@/hooks/wallets/useWallet'
-import * as walletHooks from '@/hooks/wallets/wallets'
+import * as walletHooks from '@/utils/wallets'
 import * as pending from '@/hooks/usePendingTxs'
 import * as txSender from '@/services/tx/tx-sender'
 import { useImmediatelyExecutable, useIsExecutionLoop, useTxActions, useValidateNonce } from './hooks'
