+++ src/hooks/wallets/consts.ts
@@ -1,0 +1,21 @@
+export const enum WALLET_KEYS {
+  COINBASE = 'COINBASE',
+  INJECTED = 'INJECTED',
+  KEYSTONE = 'KEYSTONE',
+  LEDGER = 'LEDGER',
+  PAIRING = 'PAIRING',
+  TREZOR = 'TREZOR',
+  WALLETCONNECT = 'WALLETCONNECT',
+  TALLYHO = 'TALLYHO',
+}
+
+export const CGW_NAMES: { [key in WALLET_KEYS]: string | undefined } = {
+  [WALLET_KEYS.COINBASE]: 'coinbase',
+  [WALLET_KEYS.INJECTED]: 'detectedwallet',
+  [WALLET_KEYS.KEYSTONE]: 'keystone',
+  [WALLET_KEYS.LEDGER]: 'ledger',
+  [WALLET_KEYS.PAIRING]: 'safeMobile',
+  [WALLET_KEYS.TREZOR]: 'trezor',
+  [WALLET_KEYS.WALLETCONNECT]: 'walletConnect',
+  [WALLET_KEYS.TALLYHO]: 'tally',
+}
