+++ src/utils/wallets.ts
@@ -4,6 +4,10 @@ import { PAIRING_MODULE_LABEL } from '@/services/pairing/module'
 import { E2E_WALLET_NAME } from '@/tests/e2e-wallet'
 import type { EthersError } from '@/utils/ethers-utils'
 import { ErrorCode } from '@ethersproject/logger'
+import { type ConnectedWallet } from '@/hooks/wallets/useOnboard'
+import { getWeb3ReadOnly } from '@/hooks/wallets/web3'
+import { EMPTY_DATA } from '@safe-global/safe-core-sdk/dist/src/utils/constants'
+import { WALLET_KEYS } from '@/hooks/wallets/consts'
 
 const isWCRejection = (err: Error): boolean => {
   return /rejected/.test(err?.message)
@@ -51,4 +55,22 @@ export const isWalletUnlocked = async (walletName: string): Promise<boolean> => 
   }
 
   return false
+}
+
+export const isHardwareWallet = (wallet: ConnectedWallet): boolean => {
+  return [WALLET_KEYS.LEDGER, WALLET_KEYS.TREZOR, WALLET_KEYS.KEYSTONE].includes(
+    wallet.label.toUpperCase() as WALLET_KEYS,
+  )
+}
+
+export const isSmartContractWallet = async (wallet: ConnectedWallet) => {
+  const provider = getWeb3ReadOnly()
+
+  if (!provider) {
+    throw new Error('Provider not found')
+  }
+
+  const code = await provider.getCode(wallet.address)
+
+  return code !== EMPTY_DATA
 }
